+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - [#360](https://github.com/influxdata/kapacitor/pull/360): Forking tasks by measurement in order to improve performance
 - [#386](https://github.com/influxdata/kapacitor/issues/386): Adds official Go HTTP client package.
+- [#399](https://github.com/influxdata/kapacitor/issues/399): Allow disabling of subscriptions.
 
 ### Bugfixes